algorithm myquick and radix

+++ push_swap/srcs/libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: dilopez- <dilopez-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/21 17:01:21 by dilopez-          #+#    #+#             */
-/*   Updated: 2022/05/21 18:03:44 by dilopez-         ###   ########.fr       */
+/*   Updated: 2022/05/31 08:27:14 by dilopez-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,11 +19,11 @@ int	ft_isdigit(int c)
 	return (0);
 }
 
-int	ft_atoi(const char *str)
+long int	ft_atoi(const char *str)
 {
-	int	negative;
-	int	number;
-	int	i;
+	int			negative;
+	long int	number;
+	int			i;
 
 	negative = 0;
 	number = 0;
